MFDP/6. Упаковка. MVP/app/requirements.txt
SQLAlchemy==2.0.41
pydantic-settings==2.10.1
sqlmodel==0.0.24
loguru==0.6.0
psycopg2-binary==2.9.10
fastapi[standard-no-fastapi-cloud-cli]==0.116.1
uvicorn==0.35.0
pyjwt==2.10.1
pika==1.3.2
pytest==8.4.1
joblib==1.5.2
autogluon==1.4.0
numpy==2.1.3
pandas==2.3.3